Create BofA_to_GS.py
~ I know the output method could be improved ~ 
It took 3-5 seconds to run on a Hyper-V 2008 windows server, likely because it opened the output file so many times ...
The majority of error checking was done by the script that got the BofA file ... FTP can fail in so many ways ...
This part of the script ran continuously until 2012.  Future development work addressed file transit errors.

I dunno if this format is still in use, but if it is, this will help some folks out there.  =)